2018 6 6

+++ src/main/java/com/nieyue/service/RoleService.java
@@ -7,5 +7,5 @@ import com.nieyue.bean.Role;
  * @author yy
  *
  */
-public interface RoleService extends BaseService<Role, Integer>{
+public interface RoleService extends BaseService<Role, Long>{
 }
