+++ src/main/java/com/nieyue/service/impl/RoleServiceImpl.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Service;
 import com.nieyue.bean.Role;
 import com.nieyue.service.RoleService;
 @Service
-public class RoleServiceImpl extends BaseServiceImpl<Role,Integer> implements RoleService{
+public class RoleServiceImpl extends BaseServiceImpl<Role,Long> implements RoleService{
 	
 }
